chore(grid): rename and move functions, move types
