+++ packages/paste-core/utilities/grid/src/index.tsx
@@ -1,2 +1,3 @@
-export * from './Grid';
-export * from './Column';
+export {Grid} from './Grid';
+export {Column} from './Column';
+export {GridProps, ColumnProps} from './types';
